requirements.txt
boto3==1.37.4
botocore==1.37.4
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.7.0
idna==3.10
jmespath==1.0.1
pymongo==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.11.3
six==1.17.0
urllib3==2.3.0
